requirements-github.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
gunicorn==21.2.0
cryptography==41.0.7
validators==0.22.0
acme==2.7.4
certbot==2.7.4
josepy==1.14.0
dnspython==2.4.2
psycopg2-binary==2.9.9
werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0